samples/python/agents/adobe_split/pyproject.toml
[project]
name = "a2a-pdf-split"
version = "0.1.0"
description = "Split PDF"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.6,<0.3.0",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "mcp>=1.9.3",
    "google-adk>=1.2.1",
    "protobuf>=6.31.1",
    "pdfservices-sdk==4.1.1",
    "google-adk>=1.0.0",
    "google-genai>=1.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
